Separated config into its own package
