+++ babel.config.js
@@ -30,57 +30,80 @@ module.exports = api => {
     }
     return settings;
   }
+
+  const nodeTargets = {
+    node: '6',
+  };
+
+  const browserTargets = {
+    browsers: [
+      'last 2 Chrome versions',
+      'last 2 ChromeAndroid versions',
+      'last 2 Firefox versions',
+      'Firefox ESR',
+      'last 2 Edge versions',
+      'last 2 Safari version',
+      'last 2 iOS version',
+    ],
+  };
+
+  /**
+   * By default, packages need to support both node and browsers
+   */
   const defaultSettings = createBabelSettings({
     envSettings: {
+      modules: env === 'es6-modules' ? false : undefined,
       targets: {
-        node: 'current',
+        ...nodeTargets,
+        ...browserTargets,
       },
     },
   });
-  const overrides = [];
+
+  /**
+   * Note that tests run in node so the target needs be noed for test envs
+   */
+  const frontendTargets = env === 'test' ? nodeTargets : browserTargets;
+
+  const storyshotsSettings = createBabelSettings({
+    envSettings: {
+      targets: {
+        ...frontendTargets,
+      },
+    },
+  });
+
+  const frontendSettings = createBabelSettings({
+    envSettings: {
+      // Webpack handles modules for us (but not in test envs!)
+      modules: env !== 'test' ? false : undefined,
+      targets: {
+        ...frontendTargets,
+      },
+    },
+  });
+
+  const overrides = [
+    {
+      test: [
+        join(__dirname, 'packages/frontend/.storybook'),
+        join(__dirname, 'packages/frontend/test/snapshots/storyshots.test.tsx'),
+      ],
+      ...storyshotsSettings,
+    },
+    {
+      test: join(__dirname, 'packages/frontend/src'),
+      ...frontendSettings,
+    },
+  ];
+
   if (env === 'test') {
     /*
       This only exists because stories are found via require.context(...)
       which is only available when code run through webpack. Snapshot tests
       aren't run through Webpack.
     */
-    const storyshotsSettings = createBabelSettings({
-      envSettings: {
-        targets: {
-          node: 'current',
-        },
-      },
-    });
     storyshotsSettings.plugins.push('require-context-hook');
-    overrides.push({
-      test: [
-        join(__dirname, 'packages/frontend/.storybook'),
-        join(__dirname, 'packages/frontend/test/snapshots/storyshots.test.tsx'),
-      ],
-      ...storyshotsSettings,
-    });
-  } else {
-    // Outside of the test env, we to make the bundle compatible with browsers instead of node
-    const frontendSettings = createBabelSettings({
-      envSettings: {
-        modules: false,
-        targets: {
-          browsers: [
-            'last 2 Chrome versions',
-            'last 2 ChromeAndroid versions',
-            'last 2 Firefox versions',
-            'Firefox ESR',
-            'last 2 Edge versions',
-            'last 2 Safari version',
-            'last 2 iOS version',
-          ],
-        },
-      },
-    });
-    overrides.push({
-      test: join(__dirname, 'packages/frontend/src'),
-      ...frontendSettings,
-    });
   }
   const config = {
     ...defaultSettings,
