+++ packages/config/jest.config.js
@@ -1,0 +1,1 @@
+module.exports = require('@openapi-platform/build-util').jest.settings();
