+++ packages/config/package.json
@@ -1,0 +1,16 @@
+{
+  "name": "@openapi-platform/config",
+  "description": "Config parser, let's you read openapi-platform.config.zzz config files",
+  "version": "1.0.0-alpha.0",
+  "main": "./lib/index.js",
+  "types": "./src/index.tsx",
+  "license": "Apache-2",
+  "private": false,
+  "sideEffects": false,
+  "dependencies": {
+    "js-yaml": "^3.12.0",
+    "json5": "^2.0.1",
+    "convict": "^4.3.2",
+    "fs-extra": "^7.0.0"
+  }
+}
