+++ packages/server/package.json
@@ -19,16 +19,12 @@
     "@types/express": "^4.16.0",
     "@types/sequelize": "^4.27.23",
     "body-parser": "^1.18.2",
-    "convict": "^4.3.2",
     "cors": "^2.8.4",
     "express": "^4.16.3",
     "feathers-memory": "^2.1.1",
     "feathers-sequelize": "^3.1.2",
     "feathers-swagger": "^0.7.1",
     "joi": "^13.6.0",
-    "js-yaml": "^3.12.0",
-    "json5": "^2.0.1",
-    "node-config": "^0.0.2",
     "pg": "^7.4.3",
     "pg-hstore": "^2.3.2",
     "sequelize": "^4.37.6",
