+++ packages/server/src/config.tsx
@@ -1,0 +1,2 @@
+import { readConfig } from '@openapi-platform/config';
+export const config = readConfig();
