--- packages/server/src/config/index.tsx
@@ -1,3 +1,0 @@
-import { readConfig } from './readConfig';
-// TODO: We can provide more ways to specify configs in the future
-export const config = readConfig();
