+++ packages/server/test/integration/server.test.tsx
@@ -1,7 +1,7 @@
 import { SdkConfig, BuildStatus } from '@openapi-platform/model';
 import { Spec } from '@openapi-platform/model';
 import { createServer } from '../../src/createServer';
-jest.mock('../../src/config/readConfig');
+jest.mock('@openapi-platform/config');
 jest.mock('@openapi-platform/logger');
 jest.mock('sequelize');
 jest.mock('feathers-sequelize');
